+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 ## [0.2.0] - 2020-01-26
 
-### Changed
 
-- Renamed `treetop.Renderer` struct to `treetop.Page`
-- Remove `treetop.NewView..` API method for creating base view with default template exec, doesn't add anything to the API
+### Breaking API Changes
+
+Taking Treetop from POC to Alpha gives me an opportunity to
+execute on a wishlist of API changes.
+
+- Rename `treetop.Renderer` type to `treetop.Page`
+- Remove `treetop.NewView..` API method for creating base view with default template exec
+- Rename `treetop.TreetopWriter` interface to `treetop.Writer` to conform to naming guidelines
+- Remove `treetop.Test`, testing recipes and resources belong elsewhere
+- Split `treetop.Writer` function into `treetop.NewPartialWriter` and `treetop.NewFragmentWriter` and remove the confusing `isPartial` flag
+
 
 #### Defining a page with views